Use shared status labels in device modal
