+++ src/features/product-passport/types.ts
@@ -1,0 +1,1 @@
+export type TemplateFieldValue = string | number | boolean | string[];
