+++ src/features/product-passport/utils/date.ts
@@ -1,0 +1,2 @@
+export const formatPassportDateTime = (value: string) =>
+  new Date(value).toLocaleString('ru-RU', { hour12: false });
