+++ src/features/product-passport/workspace/utils.ts
@@ -1,0 +1,85 @@
+import type { DeviceModel, PassportTemplateField, ProductPassport } from '../../entities';
+import type { TemplateFieldValue } from '../types';
+import type { DeviceFormValues, TemplateFieldDraft } from './types';
+
+export const createSlug = (value: string) =>
+  value
+    .trim()
+    .toLowerCase()
+    .replace(/[^a-z0-9а-яё]+/gi, '-');
+
+export const generateTempId = () => `tmp-${Math.random().toString(36).slice(2, 10)}`;
+
+export const createBlankFieldDraft = (): TemplateFieldDraft => ({
+  id: generateTempId(),
+  label: 'Наименование узла',
+  key: 'nodeName',
+  type: 'text',
+  required: true,
+  options: '',
+});
+
+export const normalizeFieldDraft = (draft: TemplateFieldDraft): PassportTemplateField => ({
+  id: draft.id,
+  key: draft.key || createSlug(draft.label) || draft.id,
+  label: draft.label,
+  type: draft.type,
+  required: draft.required,
+  placeholder: draft.placeholder,
+  defaultValue: draft.type === 'number' ? Number(draft.defaultValue) : draft.defaultValue,
+  options:
+    draft.type === 'select'
+      ? draft.options
+          .split(',')
+          .map(item => item.trim())
+          .filter(Boolean)
+          .map(item => ({ label: item, value: createSlug(item) }))
+      : undefined,
+});
+
+export const createDefaultDeviceFormValues = (models: DeviceModel[]): DeviceFormValues => ({
+  assetTag: '',
+  deviceModelId: models[0]?.id ?? '',
+  serialNumber: '',
+  ipAddress: '',
+  location: '',
+  owner: '',
+  status: 'in_service',
+  historyNote: '',
+});
+
+export const getMissingRequired = (
+  schema: PassportTemplateField[],
+  values: Record<string, TemplateFieldValue>,
+) =>
+  schema
+    .filter(field => field.required)
+    .filter(field => {
+      const value = values[field.key];
+      if (field.type === 'checkbox') {
+        return value !== true;
+      }
+      if (Array.isArray(value)) {
+        return value.length === 0;
+      }
+      return value === undefined || value === '';
+    })
+    .map(field => field.label);
+
+export const buildDetailDefaults = (passport: ProductPassport): Record<string, TemplateFieldValue> => {
+  const defaults: Record<string, TemplateFieldValue> = {};
+  passport.schema.forEach(field => {
+    const value = passport.fieldValues[field.key];
+    if (value !== undefined) {
+      defaults[field.key] = value;
+      return;
+    }
+    if (field.defaultValue !== undefined) {
+      defaults[field.key] = field.defaultValue;
+      return;
+    }
+    defaults[field.key] = field.type === 'checkbox' ? false : '';
+  });
+  return defaults;
+};
+
